fix bulk vis signal
