+++ firstvoices/backend/models/signals/__init__.py
@@ -1,1 +1,2 @@
 from .mtd_signals import *  # noqa F401
+from .visibility_signals import *  # noqa F401, F403
